third_party/WebKit/LayoutTests/external/wpt/css/cssom/cssom-fontfacerule-expected.txt

This is a testharness.js-based test.
FAIL CSSStyleDeclaration values are represented within CSSFontFaceRule assert_equals: expected "url(http://foo/bar/font.ttf)" but got "url(\"http://foo/bar/font.ttf\")"
Harness: the test ran to completion.
